Add recipes to install an array of formulas/casks
This operates similar to Matt Ray's "packages" cookbook. A list of
formulas can be created as an array, and the `install_formulas` recipe
will iterate over it and install each formula as a brew package.
Similarly, a list of casks can be specified and installed using the
`install_casks` recipe.

I used "formulas" instead of "formulae" as this pertains to "a list of
ingredients for or constituents of something", rather than of a
mathematical relationship, to use the OS X Dictionary.app's
definition. As this is for an OS X-related cookbook, I thought it
appropos to use such definition, as words matter.