Implement "removal" of provider under Chef 12
This is contingent on https://github.com/opscode/chef/pull/1921/ -
whereby the homebrew package provider is moved to Chef and set as the
default provider for `package` resources on OS X.

The change introduced in this commit attempts to maintain
compatibility seamlessly, so that users who haven't upgraded to Chef
12 will still have "homebrew as default" functionality that this
cookbook always provided.

Testing the change was done successfully with test kitchen, as output
in this gist:

https://gist.github.com/jtimberman/dcc4434b019312a83e6e

Unfortunately, the test itself requires two things:

1. Building a custom base box for OS X
2. Installing XCode, in order to install Chef 12 from the homebrew
   branch in the PR referenced above
3. Using an outstanding test-kitchen pull-request to even install Chef
   12 from github, as we don't have an omnibus package with that

This makes it awkward for other people to test. However, this cookbook
didn't exactly have tests before either.