Don't upgrade casks when an older version already is installed
`brew cask list <name>` will return with a non-zero exit code if
the installed cask version is not the latest. This breaks idempotence
of this resource when a newer version of a given cask available.
`brew cask list` will include older versions of casks so grepping
that list will give us an actual result if it's installed or not.