Stub Chef12HomebrewUser#find_homebrew_uid
The Travis build is running on a Linux VM, which does not have Homebrew.
When trying to figure out the Homebrew user, Chef12HomebrewUser requires
Homebrew to be installed, so the method fails. Stubbing it with the
current user's id allows the test to pass.

An alternate workaround is to modify .travis.yml to specify objective-c
as the language, which will cause Travis to run specs in an OS X VM with
Homebrew pre-installed.